requirements.txt
pyacoustid
musicbrainzngs
pymongo
mutagen
musicbrainzngs
ruff